Put generating code back in template/template.py
